created separate src folder for processing the data and fixed bug when graphing two separate counties with the same name
